pyproject.toml
[tool.poetry]
name = "distributed-speculative-inference"
version = "0.1.0"
description = "Distributed Speculative Inference of LLMs"
authors = ["Nadav Timor <nadav.timor@weizmann.ac.il>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "dsi"}]

[tool.poetry.dependencies]
python = "^3.10.14"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
pandas = "^2.2.2"
hydra-core = "^1.3.2"
pydantic = "^2.7.4"
ray = "^2.31.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pytest = "^8.2.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"